+++ source/server.ts
@@ -9,7 +9,8 @@ import { parse, resolve } from 'url';
 
 import { GetCacheEntry, GetCacheFunction, i18n } from './common';
 import './global';
-import { JSONObject, set, getAddCachedTranslationsJS } from './utils';
+import { getAddCachedTranslationsJS } from './getJS';
+import { JSONObject, set } from './utils';
 
 i18n.setOptions({ hostUrl: Meteor.absoluteUrl() });
 
@@ -36,7 +37,7 @@ function getJS(locale: string, namespace?: string, isBefore?: boolean) {
 
   return isBefore
     ? `var w=this||window;w.__uniI18nPre=w.__uniI18nPre||{};w.__uniI18nPre['${locale}${
-        namespace && typeof namespace === 'string' ? `.${namespace}` : ''
+        namespace ? `.${namespace}` : ''
       }'] = ${json}`
     : getAddCachedTranslationsJS(locale, json, namespace);
 }
