+++ source/utils.ts
@@ -38,13 +38,3 @@ export function set(object: UnknownRecord, path: string, value: unknown) {
 
   object[last] = value;
 }
-
-export function getAddCachedTranslationsJS(
-  locale: string,
-  data: string,
-  namespace?: string,
-) {
-  return `(Package['universe:i18n'].i18n).addTranslations('${locale}', ${
-    namespace ? `'${namespace}', ` : ''
-  }${data}),Package['universe:i18n'].i18n._ts = Math.max(Package['universe:i18n'].i18n._ts, ${Date.now()});`;
-}
